FR #557 Imported checks have contribution type "merkle"
This is also a showcase for the new wmf_civicrm message
import.  I've replaced the stomp queueing with a direct
call to import the contribution.

Made a few other minor changes to use our new apis.

Note that running this import on the same file twice will
create duplicated contacts and contributions.

There is a schema migration to add the 'Merkle' Contribution
Type.