(FR #557) fix duplicate checking
This is cheesy, but use knowledge about the trxn_id which will
eventually be generated to perform our duplicate check while
importing.

I would prefer not doing this silent check, instead throwing a
duplicate contribution exception and explicitly handling that
in the calling function, but that sounds like more work.

FU r1991