Mark the donation as successful earlier to avoid limbo states
Previously, upon a successful donation, we would first insert the
donation record into the Civi database and do thank you e-mails and all
that stuff, and all the way at the end we would mark this month's
payment as successful and reschedule it for next month.

This had the potential to get subscriptions stuck in limbo if the
insertion into civi failed (see Mingle #560), so instead, make recording
the success the *first* thing we do.