(FR #619) db switcher cleans up when finished
The db switcher was not respecting the active db.  Now, each db switch
pushes the previously active db onto a stack and restores on pop() or
when the db switcher object goes out of scope.

Also corrected uses of {table_name} on the civi db, which definitely
doesn't use drupal table prefixes.