profiling test for queue2civicrm
HOWTO: phpunit tests/performance.test

defaults are to connect to localhost, generate 1k random transactions,
push to the "test-donations" queue, and consume.  turn on xdebug profiling
before bothering with this test.

running "drush qc" as a shell call has the happy effect of creating a
callgrind output file independent of the test framework.