(FR #1097) Fix recurring text in Thank-you
Also fixes text rendering, which was broken in several ways.

This patch abandons the templates/txt/**, in favor of on-the-fly
html2txt conversion.  Now it is possible to include a link from
within the unsubscribe function, and the recurring transclusion.

The \t at the beginning of paragraphs is eliminated.

Twig cache files are stored under the Drupal sites/default/files dir.

Change-Id: I6a2b1b40d00ee6924da07f391d7972cc9f6f8783