Just pass utm_campaign to direct_mail_appeal
Anne Gomez requested that it be easier to setup campaigns for email
appeals. The current system you had to modify wmf_civicrm.module to add
a new entry to the switch statement for utm_campaign. Now it will just
pass utm_campaign to direct_mail_appeal which corresponds to an option
value for the custom field 'Direct Mail Appeal'. So you need to create
that option value with a name that exactly matches the utm_campaign
field that will get recorded with the original donation.

Change-Id: Idade44b6e022d6c907df30af905bda5d387c6390