Handle dates more carefully in recurring messages
The message timestamp on subscription signup messages was being set incorrectly
to the current time or the payment date.  This wouldn't have corrupted data,
because the timestamp is not stored nor used for currency conversion, but it
was a problem for tests where the currency is mocked by exact time.

Change-Id: I04fe4365d8933d714b96a13af07ad864c7017897