Generic import specification
* Define new file format, WmfImportFile, and require that columns are present.
This format will become the default in a future commit, and whenever we have
control over reporting, we should be using this Individual or Organization
template for our data.  See the Excel cell comments for column documentation.
* Require Original Currency, Original Amount, and Received Date columns.
Additional validations are imposed by the wmf_civicrm module and depend on the
record type.
* The Drush "import-checks" command works again, and takes an optional --type
parameter that can be used to import formats other than the default.

Bug: T88836
Change-Id: I7ed90412a234809e440ee56e5b9d7d9534327871