Add new Location Types.
For the latest data import, we need to add and update a Location Type. We
haven't added these through code before, so I added a couple of utility
functions to take care of it. It's a bit annoying because the Civi API
doesn't currently have a way to manipulate Location Types.

Change-Id: I68dc4da0137ce4ead2019cc4b6ecd771e27d42c2