T99836 Civi46 fix ensure_prefix & ensure_suffix for 4.6
The functions that ensure the prefixes & suffixes exist weren't working. I think
the issue was around the value field being set. It seems to need
to be left empty so that the api will give it an integer key.

(We should check that all individual_prefix & individual_suffix items
have numeric 'value' fields in prod)

Note that this fix also changes the ensure preferred language function
to set the value to the 2-char language code & the name to the xx_XX version
which is consistent with others in the DB. Tht guts of this function is now
shared with the suffix & prefix functions

Change-Id: Ic80b9479c3fa8df0eac0c96a8781aca850ef52cf