Civi46 adjust function to get tests passing
I'm not 100% sure this is the fix - although it 'started working' & removes uncessary code I think the issue might be existing data in the DB.

To clarify, there needs to be an option_value in the civicrm_option_value table in the 'used_for' group for 'Contributions' but the value field MUST be civicm_contribution NOT an integer. I'm not sure if this value pre-exists & needs updating or does not exist in prod. I updated upgrade hooks to add it if not existing

Note that I had previously started to add the last upgrade hook & obviously didn't complete it leaving an embarassing duplication of the last hook! So, this doesn't remove an update.

I've started numbering the update hooks with a gap to allow for upstream updates

BUG: T99836

Change-Id: I39750d0f177ea5d7ad95fbc3b550ae63ac2f5ce4