Campaign option values must exist
Civi validates that "Select" field data is a known option value, so we
play along here.

Note that this changes the way we handle unknown campaigns, and now an
exception will be thrown during message import.  The working theory is
that this case will be rare and does indicate something pathological.

Bug: T99838
Change-Id: Id3f400474b6d71a2466d52397d2c46ae746215d1