Civi46 Update report urls to open as criteria
Issue CRM-17225 We can't really change the meaning of the core urls to show without
rendering results  we have to create an option because otherwise
we'll never get it upstreamed. Giant Rabbit did add a critieria to the output
but they also removed the code that stopped reset being treated like force
which meant a behaviour change which would have been mandatory for all sites
and hence was a problem to upstream. I have been discussing with Coleman and
although this may not be the final cut this change to the urls is what I am
currently getting him to review too.

Basically if output=critieria is in the url the rows won't be rendered until you
take an action. Otherwise they will. For new reports you can choose at the point of
creation, this alters all existing reports

Commits in the CiviCRM repo make the output=criteria

BUG: T99837
Change-Id: Ie733ae9f4becba160b3a4d39cc0a2a4a0f988b3c