Add script to fill missing appeal data
This data was missed due to a patch that caused it to never validate in early Nov and
potentially  also due to not all desired campaigns being configured. The function itself could be rerun after enabling a new option and could potentially be called from a hook called when options are enabled.

Bug: T121284:
Change-Id: Ieebfdbf78f3837824bfc3e6e2b2c5ac37b02645a