Fix update script not to fall over on missing check_number
During the last round of QA escaping for check_number was re-added. I just retested on staging and
it needs the COALESCE change to ensure that the value is not NULL as the escaping function expects a string.

The change to the form of the insert query was one suggested before that I made this time since I was retesting anyway

Bug: T124979
Change-Id: I5c3d2a04a995704f5c38b79cba9421a1dccbdcae