Fix type error in Refund logic.
The function being called in the mark-refund code to set the custom field is less tolerant than the api &
does not accept True-ish values for boolean fields. We hit an error where it passed 2 as the number of contributions this year to the custom field setValues function & it was rejected as a non boolean. This is a follow up on a recent change.

Bug: T124979
Change-Id: I3c4fcd64028158c673ef708a989da5b97c57a06b