Add rudimentary performance timing drush command
This is primarily for the purposes of comparing the time it takes us to write transactions to CiviCRM. The function uses wmf processing, but bypasses the contribution_tracking for simplicity.

The real focus is whether CiviCRM processing speed changes for the sorts of functions we do

This uses the data from https://git.wikimedia.org/blob/wikimedia%2Ffundraising%2Fdash.git/master/test_data%2FDonationQueueMessages.json but I quickly discovered that conflicts were arising over the gateway id so I added some preprocessing for that (which is outside the timed section).

I didn't think the harm in duplicating the sample data was enough to warrant clever stuff around fetching it

Sample command

drush qperf-d 20 "Staging:pre-upgrade test

Bug: T127133

Change-Id: I80ba8cbd09a22a42786da4a951a5b8aed285054b