Not using referrer from messages
We store referrer right at the start in the contribution_tracking table.
There's no reason to look for it in messages, and we don't do anything
with it if we find it.  This takes it out of test data and stops looking
for it in a couple spots.

We leave the dummy value being used to instantiate a gateway adapter,
so the DonationData doesn't try to look for a request header. (maybe fixed?)

Bug: T110564
Change-Id: I8762f03bf51bd0e3bdbe45b38f4856f305dce232