Remove code to include our custom tables.
This is actually handled fine in core now as evidenced by the fact these fields appear in the fields_in_conflict above. In addition, the only custom tables we have that do not fit the
naming scheme are wmf_donor - which is deliberately excluded and wmf_contribution_extra which
does not extend contacts and does not need to be included.

Bug: T133625
Change-Id: I604b9967fdafc46fb2949f8f80366efac9c7cfd0