Alter logic for calculating wmf_donor fields
This change causes is to be calculated prior to the main contact merge, which means there
is only one mysql transaction. This will potentially have speed benefits but it will also keep
 the mysql log more readable.

The only potential gotcha I can see is that is assumes the contributions will be merged.
It's possible in the UI to choose NOT to merge the contributions but I can't think of a valid WMF use
for not merging contributions.

Bug: T133625

Change-Id: I18f2a72b0b6c351a34b3956f00ce3f81c9ce6570