Merge payments init messages
This makes it possible for a second message with the same ct_id and
order_id to update the row in payments_initial. This must be deployed
before any change to send extra messages.

As with the payments_fraud update, validates all required fields on
insert but allows updating with just the key fields.

Bug: T137590
Change-Id: I3d7395a0a97426d08a3fb7a969b561e2b5c8360a