Merge 'donor comments' custom field into note
We've got a table that contains just a blank value for the vast majority
of contributions. The last time an actual donor comment was added was in
September 2012. Newer entries are better stored in the 'note' column of
the main contribution table, and older entries were generally copied there.

After running this script, we can disable the custom field via the UI.

Bug: T142747
Change-Id: I96e0f201ff5842139fe31bba35803ca4faca4454