Mirror messages to new queues in GC audit
When the SmashPig config has verified-new and refund-new config keys,
the GC audit parser will send copies of queue messages there.

Could have done this with a MultiQueueWriter, but that would have
mutated the source fields and required transforming the body into
a DonationInterfaceMessage. This seems like the smallest code change
that will accomplish the needful.

Bug: T141484
Change-Id: I25ef9c8193276421c933c26eaf68e06003ae9982