Fix for form merges not retaining the right email address.
With batch merges the retained contact has the lower contact ID, with form merge the higher.
So, batch merges were working correctly despite an error in the alterLocationMergeData which caused
lastDonorID to never equal  mainID (ie the lower ID)  - in the batch merge they would not equal anyway unless
a convoluted set of merges had taken place. This would also only kick in on non-major gifts.

With the form however, it was incorrectly adjusting the address due to the miscalculation.

On adding tests for the batch merge scenario I hit a few cases I had to skip. I would be
good to fix these but they are a little artificial because the batch merge & form
merge should avoid them anyway

Bug: TT143236

Change-Id: I02f89630bc5eb40293814c8757a1c2c702ca3bbc