Only store the full country code if it is on our valid list.
If the country code is not on the list then prefer the 2 digit code. This means 'en' is better than 'en_NO
(Norwegian English). Separately we will look to remove the invalid options we already have in our DB.

If neither are valid store the full thing to make it easier to find them

Bug: T96410
Change-Id: Ie7e8a958c2496a738ce738dd827b7ed04b2d9fbd