Switch receipt to use original amount & currency if provided.
I'm not loving this as it turns out the custom group label & custom field labels (not machine names)
are assigned to the template. This makes it a little brittle as we could change those. However,
it does resolve the immediate problem.

I am inclined to use it for now but raise a ticket to do a second round of this to make it more
robust (perhaps with a token, or review what is assigned to the template).

BUG: T137496

Change-Id: I22858e972d240fa58b3f3d30208bbc183a5f2406