Update trigger mysql
Updating the trigger mysql to reflect new prospect fields. Once committed I will get Jeff to
run on live to start tracking.

Also, caught up in this is a new benefactor field which was created today it seems.

There are also changes to the contribution_page table, dashboard tables, price set tables
& pledge block. The correct mysql trigger script was run after the last update
but the tidy up task of updating the committed code didn't happen.

I also note a custom group has been removed. Presumably the trigger still exists for
civicrm_value_1_note_11 but I don't have permission to see it. Removing fields
without removing the trigger would normally cause lots of errors. I'm not
quite sure whether the trigger is gone or whether we escaped errors because
the whole table is gone & hence there are no updates to it. I left the drop commands in the
trigger sql in case they are there.

Bug: T147985
Bug: T142624
Change-Id: I8714fbcfbc3375e5bd4237c311fd826a23b63efb