Set gift field on import for contributions over 999.
Major Gifts are most concerned about this for $1000 and over, so only do those in order to
meet their needs without bloat.

Note that the fields in the data set have defaults but if nothing is inserted into ANY field
in the set then the default will not kick in as it is at the mysql level.

Bug: T138361
Change-Id: I4cbc5bec3e31edfd8eba3e24ac6a4bcc8ff8c1b8