Add option to download file of skipped, ignored, and error from upload.
The url to download the csv files only has minimal protection - ie. it is restricted by permission
but within that permission some could theoretically fish for files they did not create. How much of
a concern is that?

Also, my feeling is that it would be worth when we abort copying the errors and all remaining
lines (and possibly skipped & ignored) into a 'not-imported' csv as I feel that would be easier
for the users. There is some extra processing and files created, but I think it is probably warranted.

Lastly, needs some unit tests

Bug: T88460
Change-Id: I01c694f9ef225376cac49c6cd85aa1569bbda10b