Add extension with screen to do data entry on unsubscribing emails.
Although the request was to restore an unsubscribe link I could not determine what had been there previously to restore.

I opted in the end to create a screen to do the unsubscribe in a CiviCRM extension (which we can publish & share).

Currently the extension is on my git repo - https://github.com/eileenmcnaughton/org.wikimedia.unsubscribeemail

Not sure if it is important to move to wmf repo or not.

Michael has tested this on staging and says it will allow them to do sets of unsubscribes more efficiently.

Note that to install this it is necessary to:
1) set the directory path for CiviCRM extensions
2) set the url path for CiviCRM extension (for ajax reasons)
3) enable the extension
4) add the link to the page to the drupal block (I focussed on this rather than a menu entry in CiviCRM
 as it seemed to reflect DS flow.

These things can probably all be done by update script
as a follow up.

Bug: T147571
Change-Id: I321cc684d5cd4ce70a53d1b32b7094d0e6c1ed7b