Alternate fix for reconciliation report doubling some rows.
The previous fix was to filter fee rows out of the join. That works. But I wanted to trial
a bit more and to do so it seemed really useful to be able to see the total in the original
currency. Once I tried adding that it showed that the rows were still being duplicated,
only the value was blank for the 'wrong' rows. I feel replacing the left join with a WHERE
might save confusion later. Also, I think the option of original_amount total seems useful.

Bug: T152238
Change-Id: I63e2240da371bd18cb4bc37c9424365cbc27968b