Fix unsubscribe screen to hide results when data changes.
Elliot pointed out in QA that the screen was updating the name but not the results.

I think this could be quite confusing in practice so I changed it to hide the results
if the data changes.

I also revisted a couple of sources and decided I was not using the best way to reference
the data entered based on the core example
https://github.com/totten/civix/blob/master/src/CRM/CivixBundle/Resources/views/Code/angular-page.html.php#L25

and the advice here to use an object
https://github.com/angular/angular.js/wiki/Understanding-Scopes

Bug: T147571
Change-Id: Id9a2cce92a1892024cc0eab84651543ebdf1c9e1