Create & use location update function in the scenarios where we are updating.
The update function will create or update (it uses the replace api) and is slightly more expensive
as it does a look up. I have only switched to it in the one place where we have identified that we
should be updating rather than creating duplicate addresses due to this small penalty.

Extend test to ensure double addresses not created.

Rename location_update functions to indicate they are insert.

These functions create a duplicate address if called when they should be updating

Bug: T152475
Change-Id: Ia21d771b94228267c54bed4fd0de73ef4fc99dcc