Add datachecks extension.
This extension provides a drush action to find contacts with no primary address and
one to fix them:

drush cvapi Data.check check=PrimaryLocation
drush cvapi Data.fix check=PrimaryLocation

There are some instances of this currently on live so the thinking would be to run
the fix once & then scheduled the check about once a week or so to make sure
it stays fixed (since we've had a few instances of this.

The format is intended to allow other checks & fixes to be added (there is work towards
fixing the duplicate locations in it already per Bug: T152474) and things like
munged names are also candidates, and for the extension & fixes to be shareable
with the community.

I think at the moment running the check script via jenkins will probably not trigger an email
as the output is 'info-only' variety so possibly an extra drush wrapper will be needed.

Currently on staging the DB is clean of PrimaryLocation probs but running
drush cvapi Data.check

also shows some DuplicateLocation issues.

Bug: T145873
Change-Id: I3fcdb3ce19747ed2023ab24c46a31392c4addb43