Remove entity tags for 'Review'.
Cleaning these up as stakeholders have confirmed they are not using the automatic tagging.

Should we remove the tag itself? I have left for now as it seems like exactly the
sort of thing one would use tags for, now that the DB is not full of contacts tagged
for review but could be convinced to remove it.

Bug: T118904
Change-Id: Ib7847283ded3ff9bf975c51057f3457f12ec676f