Round off all the geo_codes to 4 digits
New geo tags are going in with 4 digits of precision, while the
original tagging had something like 6. This is causing conflicts in
dedupe. Round them all off to 4 digits, which is still accurate to
within 100 meters. Since we got the geo codes from a zip code db in
the first place, the extra precision was illusory. We don't need
creepy precision to plan donor events.

Bug: T155677
Change-Id: I5b2211919adc0c010c7867766f97a6a3075f2390