Pass more error details to calling function.
By using the civicrm_api3 with debug=1 & passing ->getExtraParams() to the wmfException
we can make more information about the error available higher up the chain. These are the places
where I had to do it to debug the test issue in Benevity.

In testing I also stored the exception data on the class & accessed it from the tests but
I'm not adding that at this stage because I think it could be memory intensive

Change-Id: I27d577eeb5c385996f4bd126b6ae8067a1af271a