Fix Benevity import to only soft credit one way.
Per comment from Leanne:
'I checked in with Rosie and she agreed with me that the soft credit to the company isn't necessary for donors' gifts. Since we're going to have the company reflected in the Employer field and the relationship, that will be the best indicator. So to confirm, we'd just want the donor soft credited for the matching gift, but not vice versa.'

Bug: T115044

Change-Id: I873f5120fa3f4a3f7c63eb5dc1c476b43a837ea6