Benevity import: ensure that the individual is soft credited when there is no individual part of the gift.
In this scenario the contact is created purely for the purposes of adding the matching gift.

This commit adds a test for this. In order to make the test pass I had to address some enotices.

Oddly the do_not_trade flag was one of these. That should have been resolved on T137588

Bug: T115044

Change-Id: I61d04966d6707ea8775a36bc2161ae11cc392d6e