Rerun the geocoding.
The only thing that seems odd about the geocoding function is the lack of brackets in the
for for  < 210 * . Perhaps it was in the running. Re-trying this on staging.

Staging showed that despite the LEFT() we were seeing this error:  [nativecode=1406 ** Data too long for column 'zip' at row 76141]"

The relevant geocode is 9813203 and is probably mis-countried as the US.

I think it is OK, verging on good to put in the entire code, as it will result in the geocode being accurately updated to NULL

Bug: T158271
Change-Id: I1f1ea9870d5f9ff8da0283557e80481291c68fa7