Benevity loosening, handle multiple contact use in the past.
Our last situation (hopefully) seems to be a scenario where previous Benevity enterers have
spread their love over a range of similar contact records. In this case we can't choose the
'right' one, because probably they should just be merged.

Major Gifts were not very keen on doing any manual tidy up on them so alternative
solution is to create a new one, but after the first new contribution under the new
system we should be able to differentiate & prioritise this new one as it should be
the only one with a relationship. So, do that.

If more than one has a relationship then, give up, propagate new ones each time forever.

Hopefully that won't happy, but that seems to be preferred to manual intervention as
an overall philosophy.

Bug: T115044
Change-Id: I4e49774d0944ad16667f159b676aa191fad1a381