Further removal of class_api3 call in favour of civicrm_api3
class_api is causing test failures when the updated mysqli driver tries to free memory

This replacement promotes the use of wmf_civicrm_get_option() as a cached methodology
to retrieve an option for any field. There is actually caching at the api level
so the extra caching here is probably of marginal benefit, but I am replacing a statically cached call.

Change-Id: I2b78783f87a5f352b85b3adce2dbd8f203682136