Actually cache fallback language templates
In the previous code, if there was an en-br attempt followed by an
en-za attempt, the 'en' template would not be cached under en-za.
Cache after the en-br mail:
[ 'en' => 'en template', 'en-br' => 'en template' ]
Then on the en-za attempt, it attempts to load en-za and fails, falls
back to en, and immediately returns the cached en template without
caching it under en-za.

Bug: T170435
Change-Id: I68996e2ba849e31a71df59c5782fac22f69d95b4