Update Omnimail jobs to support storing settings about multiple jobs.
Omnimail jobs store information about where the job is up to. This change
permits an input parameter 'job_suffix' to denote a job,
allowing multiple jobs of the same type to be tracked.

This might be used if a slow back-fill job is running off-peak
(ahem) due to possibly a mistake cause some data to have been missed
.....
without interfering with the up-to-date job.

Bug: T176255
Change-Id: If0a709b9e9cec7e3276d231e75cca49d2909d4cb