Remove attempt to set sort_name & display_name.
On testing I found that what we were setting was being ignored (otherwise the cleanups
would not have flowed through).

I have done an upstream PR to address length handling in core
https://github.com/civicrm/civicrm-core/pull/11063

I don't think it's pressing for us to address long_name handling (ie.
we can suck that PR in now or wait for the next update) but
removing these 2 lines gets rid of the misleading impression that they
have an effect

Change-Id: I0be0a8883e19378019adf29f9b4cb76add4e812e