Always delete 'worse' address
Previous comment said 'leave it for Civi to deal with', but the hook
is run right before the last thing that happens with addresses. Looks
like we have to deal with it here, or we'll end up with duplicate is_primary
addresses.

To recreate the issue, create two mergable contacts. The older contact should
have an address with all information filled in. The newer contact should have
a newer contribution and should have an address with only the country filled
in, matching the country of the old contact.

Bug: T178819
Change-Id: I12799a44e4e417fc7d90be6382e7915c5d292726