Tidy up dedupe rules & install dedupetools.
My thinking is to move our merge hooks etc into this extension. It contains the code
I used for analysing the dupes & now the code for extra links to dedupe

I have removed some rules & renamed one and added a new one. The one I added is
intended to be used by Major Gifts to find matches for a specific contact, casting a
wide net.

I tested it on staging and could find  a few names that were slower than others (eg. John Smith)
but none that were server-slaughtering beasts.

Bug: T177999
Change-Id: Iaa444cc7d232b0f74a19372a8d434f0adf34e4c0