Add stripe import
At this stage an exception is thrown if non-USD currencies are in use and utm_campaign column is respected
if set, ditto utm_medium with a default of MGEventEmail

There is a desire for currency support & I think that might be the case for another import too.
I also feel there is a desire for being able to set an import-wide default for a field. I've left these
out of this iteration.

Bug: T172423

Change-Id: I63a85f352caec1d695724a9b1dd1129f9a29781c