Fix checks for existing contributions on Benevity imports to check donor AND matching contributions.
Currently donor contributions that have already been imported are marked as duplicates but where the matching
contribution has been imported that is not picked up as a duplicate and results in an error.

The impact on users is that a string of errors will result in the import aborting. On the other hand, a string
of duplicates will be ignored, allowing a mostly imported csv to be re-attempted.

Bug: T181548
Change-Id: Iff464153e40c9f751c63e21b0236d552f1e1ceb9