Allow Major gift contacts with conflicting addresses to merge.
By removing the conflict we allow all the prioritisation rules developed for other
contacts to apply

Unit tests updated to reflect the fact that merge now takes place and the most recent
donor's details are now used. Comfortingly all the 'resolvable' conflict tests are
unchanged, still selecting the preferable addresses. The 'unresolvable' conflict
tests have been updated to reflect use of most recent donor detail. (a couple
of test names had to be updated to be accurate about that distinction

Bug: T181088
Change-Id: Ie78e3a517b796967976936b4fc93d2352a7aa14e