Add civitoken extension.
This extension adds a load of tokens for use in mailings. The one Major Gifts require is the
address block token, which prints out an address block per site config, without extra
lines where a field (e.g city) is blank.

Civitoken is an existing extension which is a bit of a dumping ground for random
tokens. I considered trying to port this one into core... until I saw the core code
and also trying to strip the relevant extension out into something separate, but
I opted for creating the ability to disable unloved tokens. The ones I have enabled are

address_block - formatted address block in html, suitable for MG letters
address_text  - formatted address block in text - mostly for text only emails
conditional_country - I think this is required for the above, printed country if not US
(since WMF is configured as US)
today_format_full - yes this really is missing from core - a today's date token
today_format_raw - same deal, mostly useful if some sort of code comparison is in play
so maybe not so much for us.

Bug: T184647
Change-Id: I2c6c77acf947b4a252f8708cf6cc938c6109793a