Add module_load_include to the DonationQueueTest & OrphanSlayerTest
For CI but I hit a

Fatal error: Class 'AmazonDonationMessage' not found in /Users/emcnaughton/buildkit/build/d46/sites/all/modules/queue2civicrm/tests/phpunit/DonationQueueTest.php on line 374

I think the issue is moving between deployment branch & master the registry might not hold all entries. Adding this line
eases that a little.

From here it seems this is the right way to load it - or at least ... not wrong
https://drupal.stackexchange.com/questions/861/are-files-from-custom-module-info-not-included-on-form-submits

Change-Id: Ia44131818a9bec7f1cf5356b04761de99ff41cc6