Test and band-aid for Ingenico recurring
Currently, we are losing 100% of first-time Ingenico recurring donations
that should have been coming in via the audit parser.  This is because of
the utm data comparison. All recurring CC donations get a utm payment
method (part of the utm_source field) of 'rcc'. The message format on the
wire, recurring or not, has CC payments with payment_method 'cc', so all
the recurring ones are getting booted by the utm_ consistency check.

This band-aid fixes the rcc/cc inconsistency.

FIXME: we need a real implementation of the recurring check that works
for non-initial payments.

FIXME: 'fakedb' and 'test' options are horrible

Bug: T186011
Bug: T182646
Change-Id: Iea9cac15931c7e2226987931569cd7d0ba1267c6