Add geocoder extension
This replaces our geocoding code. Still to do is to remove the original table because
I renamed it. The extension details are in the README - the approach might seem like a bit
of overkill but I wanted our extension to become the 'go-to' for gecoding with CiviCRM.

I will follow up with a blog.

Bug: T184830

Change-Id: I56b63f9176a798146b793eff633568feaf070127