Add donor totals fields
This can be done in advance of the population of them

Note this should not be deployed until AFTER the new do_not_solicit field is
added & triggers reloaded. We want the trigger reload to start logging that new field
but the NEXT trigger reload after that one will stop logging to wmf_donor.

The timing is less an issue for deployment as it is to when we regenerate triggers
(ie. once ops have reloaded triggers for do not solicit they don't need to
then do them after these fields are added, the next time they would do
the reload would be when we drop the logging on the old do_not_solicit field

Bug: T186168
Change-Id: I15a030fa18cace54054409b1741af0bb63ba36f7