Add triggers to manage yearly totals & associated custom fields
With the new custom fields I have switched to triggers to keep the values updated.

The custom data fields are calculated from the data in the DB so are effectively just
a caching mechanism. We are already using triggers for other purposes (e.g logging,
managing the contact modified date) so any downsides of triggers we are already
incurring. As this is 'deterministic' & recalculable data I don't think this
makes any trigger downsides any worse.

There are 2 places where the calculated fields were being updated, the merge hook & the post hook.

When we did DB updates or direct mysql queries were used the fields were not updated
(that's an advantage of a trigger).

These used slightly different methodologies - the post hook was drawing transaction information
into the query whereas the merge hook was relying on contribution data only.

I dug into this & I believe the reason was that in the past (& to some extent still)
we have had our own ways of recording refunds & chargebacks that are a little ... tricksy.
I know we did do some data clean up on this & there seem to be few enough records
that would miscalculate we can possibly ignore this.

My feeling is that we should get standardised on this & fix the data / incoming,
allowing us to use the simple contribution totals. Advantages of this are that the
contribution data is available at the time when the contribution insert happens
(transaction data is created later), it's the same as the contribution tab summary display,
it is a cheaper query & it doesn't break if transaction data is incorrect.

We currently have

16010 contributions with a financial type of 'refund' or chargeback'

15855  of which have negative amounts (so will calculate fine)
155 of which have positive amounts - I'm not 100% sure what they mean, but they are still
occassionally being created

(157 including all negative-seeming financial types)

Searching on negative statuses - these are excuded from the calculation. There
would be an issue if there were negative entries (ie. to be deducted from the
totals - but only 4 of them are).

There is a minor issue using contribution record & excluding by
status in terms of dates when the refund is at a later date but I think
that is not significant for us, nor are partial payments.

In terms of performance. I did a few tests & generally found the slowest query
that I could do took about 220 ms on our slowest contact but none of them took long
enough to register on a 'normal contact - ie. 0.00 s
- some details here
https://phabricator.wikimedia.org/T186165#4099487

Compared to what we already do this should be faster as we do less queries & connections.
Also, using the contribution totals over the financial trxn totals on import is faster.

Since performance seemed pretty good I chose readability of triggers & code over squeezing
out the following potential gains
1) excluding the join on contribution_extra on insert (needed on update, not so much on insert
as the data is not created yet)
2) figuring out the relevant year fields to update (esp on insert) and only calculating those
years

Deployment notes
- I am expecting to deploy this in advance of demobbing the old fields & keeping both for a little bit
- once the triggers are in place we can do an update script, e.g update records where contribution received
in 2016 AND total_2016_2017 = 0 (ie. we will skip those the triggers have populated & can do the
upgrade afterwards).

Bug: T186168
Change-Id: I61ed03d70a0886ad0edc5d12e0a4d09bd6dcd1e6